services/payments_service/pyproject.toml
[project]
name = "payments-service"
version = "0.1.0"
description = "Payments orchestration service for fintech practice stack"
requires-python = ">=3.11"
dependencies = [
    "fastapi",
    "sqlalchemy",
    "alembic",
    "asyncpg",
    "psycopg[binary]",
    "pydantic",
    "pydantic-settings",
    "redis",
    "loguru",
    "prometheus-client",
    "opentelemetry-instrumentation-fastapi",
    "opentelemetry-sdk",
    "httpx",
    "python-jose[cryptography]"
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
